+++ cpp/mrd/ismrmrd_to_mrd.cc
@@ -5,6 +5,8 @@
 #include <ismrmrd/dataset.h>
 #include <ismrmrd/meta.h>
 #include <ismrmrd/serialization_iostream.h>
+#include <ismrmrd/xml.h>
+#include <ismrmrd/version.h>
 #include <xtensor/xview.hpp>
 
 yardl::Date date_from_string(const std::string &s)
@@ -19,17 +21,6 @@ yardl::Date date_from_string(const std::string &s)
     d = yardl::Date(ld.time_since_epoch());
 #endif
     return d;
-}
-
-std::string date_to_string(const yardl::Date &d)
-{
-    std::stringstream ss;
-#if __cplusplus < 202002L
-    ss << date::format("%F", d);
-#else
-    ss << date::format("%F", date::local_days(d));
-#endif
-    return ss.str();
 }
 
 yardl::Time time_from_string(const std::string &s)
@@ -49,17 +40,6 @@ yardl::Time time_from_string(const std::string &s)
         throw std::runtime_error("invalid date format");
     }
     return t;
-};
-
-std::string time_to_string(const yardl::Time &t)
-{
-    std::stringstream ss;
-#if __cplusplus < 202002L
-    ss << date::format("%F", t);
-#else
-    ss << date::format("%F", date::local_days(t));
-#endif
-    return ss.str();
 };
 
 // Convert ISMRMRD::SubjectInformation to mrd::SubjectInformationType
@@ -116,58 +96,6 @@ mrd::SubjectInformationType convert(ISMRMRD::SubjectInformation &subjectInformat
     return s;
 }
 
-// Convert mrd::SubjectInformationType to ISMRMRD::SubjectInformation
-ISMRMRD::SubjectInformation convert(mrd::SubjectInformationType &subjectInformation)
-{
-    ISMRMRD::SubjectInformation s;
-    if (subjectInformation.patient_name)
-    {
-        s.patientName = *subjectInformation.patient_name;
-    }
-
-    if (subjectInformation.patient_weightkg)
-    {
-        s.patientWeight_kg = *subjectInformation.patient_weightkg;
-    }
-
-    if (subjectInformation.patient_heightm)
-    {
-        s.patientHeight_m = *subjectInformation.patient_heightm;
-    }
-
-    if (subjectInformation.patient_id)
-    {
-        s.patientID = *subjectInformation.patient_id;
-    }
-
-    if (subjectInformation.patient_birthdate)
-    {
-        s.patientBirthdate = date_to_string(*subjectInformation.patient_birthdate);
-    }
-
-    if (subjectInformation.patient_gender)
-    {
-        if (*subjectInformation.patient_gender == mrd::PatientGender::kF)
-        {
-            s.patientGender = "F";
-        }
-        else if (*subjectInformation.patient_gender == mrd::PatientGender::kM)
-        {
-            s.patientGender = "M";
-        }
-        else if (*subjectInformation.patient_gender == mrd::PatientGender::kO)
-        {
-            s.patientGender = "O";
-        }
-        else
-        {
-            throw std::runtime_error("Unknown gender");
-        }
-    }
-
-    return s;
-}
-
 // Convert ISMRMRD::StudyInformation to mrd::StudyInformationType
 mrd::StudyInformationType convert(ISMRMRD::StudyInformation &studyInformation)
 {
@@ -216,54 +144,6 @@ mrd::StudyInformationType convert(ISMRMRD::StudyInformation &studyInformation)
     return s;
 }
 
-// Convert mrd::StudyInformationType to ISMRMRD::StudyInformation
-ISMRMRD::StudyInformation convert(mrd::StudyInformationType &studyInformation)
-{
-    ISMRMRD::StudyInformation s;
-
-    if (studyInformation.study_date)
-    {
-        s.studyDate = date_to_string(*studyInformation.study_date);
-    }
-
-    if (studyInformation.study_time)
-    {
-        s.studyTime = time_to_string(*studyInformation.study_time);
-    }
-
-    if (studyInformation.study_id)
-    {
-        s.studyID = *studyInformation.study_id;
-    }
-
-    if (studyInformation.accession_number)
-    {
-        s.accessionNumber = *studyInformation.accession_number;
-    }
-
-    if (studyInformation.referring_physician_name)
-    {
-        s.referringPhysicianName = *studyInformation.referring_physician_name;
-    }
-
-    if (studyInformation.study_description)
-    {
-        s.studyDescription = *studyInformation.study_description;
-    }
-
-    if (studyInformation.study_instance_uid)
-    {
-        s.studyInstanceUID = *studyInformation.study_instance_uid;
-    }
-
-    if (studyInformation.body_part_examined)
-    {
-        s.bodyPartExamined = *studyInformation.body_part_examined;
-    }
-
-    return s;
-}
-
 // Convert ISMRMRD patient position string to mrd::PatientPosition
 mrd::PatientPosition patient_position_from_string(std::string &s)
 {
@@ -305,61 +185,10 @@ mrd::PatientPosition patient_position_from_string(std::string &s)
     }
 }
 
-// Convert mrd::PatientPosition to ISMRMRD patient position string
-std::string patient_position_to_string(mrd::PatientPosition &p)
-{
-    if (p == mrd::PatientPosition::kHFP)
-    {
-        return "HFP";
-    }
-    else if (p == mrd::PatientPosition::kHFS)
-    {
-        return "HFS";
-    }
-    else if (p == mrd::PatientPosition::kHFDR)
-    {
-        return "HFDR";
-    }
-    else if (p == mrd::PatientPosition::kHFDL)
-    {
-        return "HFDL";
-    }
-    else if (p == mrd::PatientPosition::kFFP)
-    {
-        return "FFP";
-    }
-    else if (p == mrd::PatientPosition::kFFS)
-    {
-        return "FFS";
-    }
-    else if (p == mrd::PatientPosition::kFFDR)
-    {
-        return "FFDR";
-    }
-    else if (p == mrd::PatientPosition::kFFDL)
-    {
-        return "FFDL";
-    }
-    else
-    {
-        throw std::runtime_error("Unknown Patient Position");
-    }
-}
-
 // Convert ISMRMRD::threeDimensionalFloat to mrd::ThreeDimensionalFloat
 mrd::ThreeDimensionalFloat convert(ISMRMRD::threeDimensionalFloat &threeDimensionalFloat)
 {
     mrd::ThreeDimensionalFloat t;
-    t.x = threeDimensionalFloat.x;
-    t.y = threeDimensionalFloat.y;
-    t.z = threeDimensionalFloat.z;
-    return t;
-}
-
-// Convert mrd::ThreeDimensionalFloat to ISMRMRD::threeDimensionalFloat
-ISMRMRD::threeDimensionalFloat convert(mrd::ThreeDimensionalFloat &threeDimensionalFloat)
-{
-    ISMRMRD::threeDimensionalFloat t;
     t.x = threeDimensionalFloat.x;
     t.y = threeDimensionalFloat.y;
     t.z = threeDimensionalFloat.z;
@@ -372,15 +201,6 @@ mrd::MeasurementDependencyType convert(ISMRMRD::MeasurementDependency &measureme
     mrd::MeasurementDependencyType m;
     m.measurement_id = measurementDependency.measurementID;
     m.dependency_type = measurementDependency.dependencyType;
-    return m;
-}
-
-// Convert mrd::MeasurementDependencyType to ISMRMRD::MeasurementDependency
-ISMRMRD::MeasurementDependency convert(mrd::MeasurementDependencyType &measurementDependency)
-{
-    ISMRMRD::MeasurementDependency m;
-    m.measurementID = measurementDependency.measurement_id;
-    m.dependencyType = measurementDependency.dependency_type;
     return m;
 }
 
@@ -459,81 +279,6 @@ mrd::MeasurementInformationType convert(ISMRMRD::MeasurementInformation &measure
     return m;
 }
 
-// Convert mrd::MeasurementInformationType to ISMRMRD::MeasurementInformation
-ISMRMRD::MeasurementInformation convert(mrd::MeasurementInformationType &measurementInformation)
-{
-    ISMRMRD::MeasurementInformation m;
-
-    if (measurementInformation.measurement_id)
-    {
-        m.measurementID = *measurementInformation.measurement_id;
-    }
-
-    if (measurementInformation.series_date)
-    {
-        m.seriesDate = date_to_string(*measurementInformation.series_date);
-    }
-
-    if (measurementInformation.series_time)
-    {
-        m.seriesTime = time_to_string(*measurementInformation.series_time);
-    }
-
-    m.patientPosition = patient_position_to_string(measurementInformation.patient_position);
-
-    if (measurementInformation.relative_table_position)
-    {
-        m.relativeTablePosition = convert(*measurementInformation.relative_table_position);
-    }
-
-    if (measurementInformation.initial_series_number)
-    {
-        m.initialSeriesNumber = *measurementInformation.initial_series_number;
-    }
-
-    if (measurementInformation.protocol_name)
-    {
-        m.protocolName = *measurementInformation.protocol_name;
-    }
-
-    if (measurementInformation.sequence_name)
-    {
-        m.sequenceName = *measurementInformation.sequence_name;
-    }
-
-    if (measurementInformation.series_description)
-    {
-        m.seriesDescription = *measurementInformation.series_description;
-    }
-
-    for (auto &dependency : measurementInformation.measurement_dependency)
-    {
-        m.measurementDependency.push_back(convert(dependency));
-    }
-
-    if (measurementInformation.series_instance_uid_root)
-    {
-        m.seriesInstanceUIDRoot = *measurementInformation.series_instance_uid_root;
-    }
-
-    if (measurementInformation.frame_of_reference_uid)
-    {
-        m.frameOfReferenceUID = *measurementInformation.frame_of_reference_uid;
-    }
-
-    if (measurementInformation.referenced_image_sequence)
-    {
-        for (auto &image : measurementInformation.referenced_image_sequence->referenced_sop_instance_uid)
-        {
-            ISMRMRD::ReferencedImageSequence referencedImage;
-            referencedImage.referencedSOPInstanceUID = image;
-            m.referencedImageSequence.push_back(referencedImage);
-        }
-    }
-
-    return m;
-}
-
 // Convert ISMRMRD::AcquisitionSystemInformation to mrd::AcquisitionSystemInformationType
 mrd::AcquisitionSystemInformationType convert(ISMRMRD::AcquisitionSystemInformation &a)
 {
@@ -598,70 +343,6 @@ mrd::AcquisitionSystemInformationType convert(ISMRMRD::AcquisitionSystemInformat
     return asi;
 }
 
-// Convert mrd::AcquisitionSystemInformationType to ISMRMRD::AcquisitionSystemInformation
-ISMRMRD::AcquisitionSystemInformation convert(mrd::AcquisitionSystemInformationType &a)
-{
-    ISMRMRD::AcquisitionSystemInformation asi;
-
-    if (a.system_vendor)
-    {
-        asi.systemVendor = *a.system_vendor;
-    }
-
-    if (a.system_model)
-    {
-        asi.systemModel = *a.system_model;
-    }
-
-    if (a.system_field_strength_t)
-    {
-        asi.systemFieldStrength_T = *a.system_field_strength_t;
-    }
-
-    if (a.relative_receiver_noise_bandwidth)
-    {
-        asi.relativeReceiverNoiseBandwidth = *a.relative_receiver_noise_bandwidth;
-    }
-
-    if (a.receiver_channels)
-    {
-        asi.receiverChannels = *a.receiver_channels;
-    }
-
-    if (a.coil_label.size() > 0)
-    {
-        for (auto &c : a.coil_label)
-        {
-            ISMRMRD::CoilLabel cl;
-            cl.coilName = c.coil_name;
-            cl.coilNumber = c.coil_number;
-            asi.coilLabel.push_back(cl);
-        }
-    }
-
-    if (a.institution_name)
-    {
-        asi.institutionName = *a.institution_name;
-    }
-
-    if (a.station_name)
-    {
-        asi.stationName = *a.station_name;
-    }
-
-    if (a.device_id)
-    {
-        asi.deviceID = *a.device_id;
-    }
-
-    if (a.device_serial_number)
-    {
-        asi.deviceSerialNumber = *a.device_serial_number;
-    }
-
-    return asi;
-}
-
 // Convert ISMRMRD::ExperimentalConditions to mrd::ExperimentalConditionsType
 mrd::ExperimentalConditionsType convert(ISMRMRD::ExperimentalConditions &e)
 {
@@ -670,28 +351,10 @@ mrd::ExperimentalConditionsType convert(ISMRMRD::ExperimentalConditions &e)
     return ec;
 }
 
-// Convert mrd::ExperimentalConditionsType to ISMRMRD::ExperimentalConditions
-ISMRMRD::ExperimentalConditions convert(mrd::ExperimentalConditionsType &e)
-{
-    ISMRMRD::ExperimentalConditions ec;
-    ec.H1resonanceFrequency_Hz = e.h1resonance_frequency_hz;
-    return ec;
-}
-
 // Convert ISMRMRD::MatrixSize to mrd::MatrixSizeType
 mrd::MatrixSizeType convert(ISMRMRD::MatrixSize &m)
 {
     mrd::MatrixSizeType matrixSize;
-    matrixSize.x = m.x;
-    matrixSize.y = m.y;
-    matrixSize.z = m.z;
-    return matrixSize;
-}
-
-// Convert mrd::MatrixSizeType to ISMRMRD::MatrixSize
-ISMRMRD::MatrixSize convert(mrd::MatrixSizeType &m)
-{
-    ISMRMRD::MatrixSize matrixSize;
     matrixSize.x = m.x;
     matrixSize.y = m.y;
     matrixSize.z = m.z;
@@ -708,16 +371,6 @@ mrd::FieldOfViewmm convert(ISMRMRD::FieldOfView_mm &f)
     return fieldOfView;
 }
 
-// Convert mrd::FieldOfViewmm to ISMRMRD::FieldOfView_mm
-ISMRMRD::FieldOfView_mm convert(mrd::FieldOfViewmm &f)
-{
-    ISMRMRD::FieldOfView_mm fieldOfView;
-    fieldOfView.x = f.x;
-    fieldOfView.y = f.y;
-    fieldOfView.z = f.z;
-    return fieldOfView;
-}
-
 // Convert ISMRMRD::EncodingSpace to mrd::EncodingSpaceType
 mrd::EncodingSpaceType convert(ISMRMRD::EncodingSpace &e)
 {
@@ -727,29 +380,10 @@ mrd::EncodingSpaceType convert(ISMRMRD::EncodingSpace &e)
     return encodingSpace;
 }
 
-// Convert mrd::EncodingSpaceType to ISMRMRD::EncodingSpace
-ISMRMRD::EncodingSpace convert(mrd::EncodingSpaceType &e)
-{
-    ISMRMRD::EncodingSpace encodingSpace;
-    encodingSpace.matrixSize = convert(e.matrix_size);
-    encodingSpace.fieldOfView_mm = convert(e.field_of_viewmm);
-    return encodingSpace;
-}
-
 // Convert ISMRMRD::Limit to mrd::LimitType
 mrd::LimitType convert(ISMRMRD::Limit &l)
 {
     mrd::LimitType limit;
-    limit.minimum = l.minimum;
-    limit.maximum = l.maximum;
-    limit.center = l.center;
-    return limit;
-}
-
-// Convert mrd::LimitType to ISMRMRD::Limit
-ISMRMRD::Limit convert(mrd::LimitType &l)
-{
-    ISMRMRD::Limit limit;
     limit.minimum = l.minimum;
     limit.maximum = l.maximum;
     limit.center = l.center;
@@ -854,117 +488,10 @@ mrd::EncodingLimitsType convert(ISMRMRD::EncodingLimits &e)
     return encodingLimits;
 }
 
-// Convert mrd::EncodingLimitsType to ISMRMRD::EncodingLimits
-ISMRMRD::EncodingLimits convert(mrd::EncodingLimitsType &encodingLimit)
-{
-    ISMRMRD::EncodingLimits el;
-
-    if (encodingLimit.kspaceencodingstep0)
-    {
-        el.kspace_encoding_step_0 = convert(*encodingLimit.kspaceencodingstep0);
-    }
-
-    if (encodingLimit.kspaceencodingstep1)
-    {
-        el.kspace_encoding_step_1 = convert(*encodingLimit.kspaceencodingstep1);
-    }
-
-    if (encodingLimit.kspaceencodingstep2)
-    {
-        el.kspace_encoding_step_2 = convert(*encodingLimit.kspaceencodingstep2);
-    }
-
-    if (encodingLimit.average)
-    {
-        el.average = convert(*encodingLimit.average);
-    }
-
-    if (encodingLimit.slice)
-    {
-        el.slice = convert(*encodingLimit.slice);
-    }
-
-    if (encodingLimit.contrast)
-    {
-        el.contrast = convert(*encodingLimit.contrast);
-    }
-
-    if (encodingLimit.phase)
-    {
-        el.phase = convert(*encodingLimit.phase);
-    }
-
-    if (encodingLimit.repetition)
-    {
-        el.repetition = convert(*encodingLimit.repetition);
-    }
-
-    if (encodingLimit.set)
-    {
-        el.set = convert(*encodingLimit.set);
-    }
-
-    if (encodingLimit.segment)
-    {
-        el.segment = convert(*encodingLimit.segment);
-    }
-
-    if (encodingLimit.user0)
-    {
-        el.user[0] = convert(*encodingLimit.user0);
-    }
-
-    if (encodingLimit.user1)
-    {
-        el.user[1] = convert(*encodingLimit.user1);
-    }
-
-    if (encodingLimit.user2)
-    {
-        el.user[2] = convert(*encodingLimit.user2);
-    }
-
-    if (encodingLimit.user3)
-    {
-        el.user[3] = convert(*encodingLimit.user3);
-    }
-
-    if (encodingLimit.user4)
-    {
-        el.user[4] = convert(*encodingLimit.user4);
-    }
-
-    if (encodingLimit.user5)
-    {
-        el.user[5] = convert(*encodingLimit.user5);
-    }
-
-    if (encodingLimit.user6)
-    {
-        el.user[6] = convert(*encodingLimit.user6);
-    }
-
-    if (encodingLimit.user7)
-    {
-        el.user[7] = convert(*encodingLimit.user7);
-    }
-
-    return el;
-}
-
 // Convert ISMRMRD::UserParameterLong to mrd::UserParameterLongType
 mrd::UserParameterLongType convert(ISMRMRD::UserParameterLong &u)
 {
     mrd::UserParameterLongType userParameterLong;
-    userParameterLong.name = u.name;
-    userParameterLong.value = u.value;
-    return userParameterLong;
-}
-
-// Convert mrd::UserParameterLongType to ISMRMRD::UserParameterLong
-ISMRMRD::UserParameterLong convert(mrd::UserParameterLongType &u)
-{
-    ISMRMRD::UserParameterLong userParameterLong;
     userParameterLong.name = u.name;
     userParameterLong.value = u.value;
     return userParameterLong;
@@ -979,28 +506,10 @@ mrd::UserParameterDoubleType convert(ISMRMRD::UserParameterDouble &u)
     return userParameterDouble;
 }
 
-// Convert mrd::UserParameterDoubleType to ISMRMRD::UserParameterDouble
-ISMRMRD::UserParameterDouble convert(mrd::UserParameterDoubleType &u)
-{
-    ISMRMRD::UserParameterDouble userParameterDouble;
-    userParameterDouble.name = u.name;
-    userParameterDouble.value = u.value;
-    return userParameterDouble;
-}
-
 // Convert ISMRMRD::UserParameterString to mrd::UserParameterStringType
 mrd::UserParameterStringType convert(ISMRMRD::UserParameterString &u)
 {
     mrd::UserParameterStringType userParameterString;
-    userParameterString.name = u.name;
-    userParameterString.value = u.value;
-    return userParameterString;
-}
-
-// Convert mrd::UserParameterStringType to ISMRMRD::UserParameterString
-ISMRMRD::UserParameterString convert(mrd::UserParameterStringType &u)
-{
-    ISMRMRD::UserParameterString userParameterString;
     userParameterString.name = u.name;
     userParameterString.value = u.value;
     return userParameterString;
@@ -1035,50 +544,12 @@ mrd::TrajectoryDescriptionType convert(ISMRMRD::TrajectoryDescription &t)
     return trajectoryDescription;
 }
 
-// Convert mrd::TrajectoryDescriptionType to ISMRMRD::TrajectoryDescription
-ISMRMRD::TrajectoryDescription convert(mrd::TrajectoryDescriptionType &t)
-{
-    ISMRMRD::TrajectoryDescription trajectoryDescription;
-    trajectoryDescription.identifier = t.identifier;
-
-    for (auto &u : t.user_parameter_long)
-    {
-        trajectoryDescription.userParameterLong.push_back(convert(u));
-    }
-
-    for (auto &u : t.user_parameter_double)
-    {
-        trajectoryDescription.userParameterDouble.push_back(convert(u));
-    }
-
-    for (auto &u : t.user_parameter_string)
-    {
-        trajectoryDescription.userParameterString.push_back(convert(u));
-    }
-
-    if (t.comment)
-    {
-        trajectoryDescription.comment = *t.comment;
-    }
-
-    return trajectoryDescription;
-}
-
 // Convert ISMRMRD::AccelerationFactor to mrd::AccelerationFactorType
 mrd::AccelerationFactorType convert(ISMRMRD::AccelerationFactor &a)
 {
     mrd::AccelerationFactorType accelerationFactor;
     accelerationFactor.kspaceencodingstep1 = a.kspace_encoding_step_1;
     accelerationFactor.kspaceencodingstep2 = a.kspace_encoding_step_2;
-    return accelerationFactor;
-}
-
-// Convert mrd::AccelerationFactorType to ISMRMRD::AccelerationFactor
-ISMRMRD::AccelerationFactor convert(mrd::AccelerationFactorType &a)
-{
-    ISMRMRD::AccelerationFactor accelerationFactor;
-    accelerationFactor.kspace_encoding_step_1 = a.kspaceencodingstep1;
-    accelerationFactor.kspace_encoding_step_2 = a.kspaceencodingstep2;
     return accelerationFactor;
 }
 
@@ -1111,26 +582,6 @@ mrd::CalibrationMode calibration_mode_from_string(std::string &m)
     }
 }
 
-// Convert mrd::CalibrationMode to ISMRMRD calibration mode string
-std::string calibration_mode_to_string(mrd::CalibrationMode &m)
-{
-    switch (m)
-    {
-    case mrd::CalibrationMode::kEmbedded:
-        return "embedded";
-    case mrd::CalibrationMode::kInterleaved:
-        return "interleaved";
-    case mrd::CalibrationMode::kSeparate:
-        return "separate";
-    case mrd::CalibrationMode::kExternal:
-        return "external";
-    case mrd::CalibrationMode::kOther:
-        return "other";
-    default:
-        throw std::runtime_error("Unknown CalibrationMode: " + std::to_string((int)m));
-    }
-}
-
 // Convert ISMRMRD interleaving dimension string to mrd::InterleavingDimension
 mrd::InterleavingDimension interleaving_dimension_from_string(std::string &s)
 {
@@ -1160,26 +611,6 @@ mrd::InterleavingDimension interleaving_dimension_from_string(std::string &s)
     }
 }
 
-// Convert mrd::InterleavingDimension to ISMRMRD interleaving dimension string
-std::string interleaving_dimension_to_string(mrd::InterleavingDimension &d)
-{
-    switch (d)
-    {
-    case mrd::InterleavingDimension::kPhase:
-        return "phase";
-    case mrd::InterleavingDimension::kRepetition:
-        return "repetition";
-    case mrd::InterleavingDimension::kContrast:
-        return "contrast";
-    case mrd::InterleavingDimension::kAverage:
-        return "average";
-    case mrd::InterleavingDimension::kOther:
-        return "other";
-    default:
-        throw std::runtime_error("Unknown InterleavingDimension: " + std::to_string((int)d));
-    }
-}
-
 // Convert ISMRMRD::MultibandSpacing to mrd::MultibandSpacingType
 mrd::MultibandSpacingType convert(ISMRMRD::MultibandSpacing &m)
 {
@@ -1187,17 +618,6 @@ mrd::MultibandSpacingType convert(ISMRMRD::MultibandSpacing &m)
     for (auto s : m.dZ)
     {
         multibandSpacing.d_z.push_back(s);
-    }
-    return multibandSpacing;
-}
-
-// Convert mrd::MultibandSpacingType to ISMRMRD::MultibandSpacing
-ISMRMRD::MultibandSpacing convert(mrd::MultibandSpacingType &m)
-{
-    ISMRMRD::MultibandSpacing multibandSpacing;
-    for (auto s : m.d_z)
-    {
-        multibandSpacing.dZ.push_back(s);
     }
     return multibandSpacing;
 }
@@ -1223,27 +643,6 @@ mrd::Calibration convert(ISMRMRD::MultibandCalibrationType &m)
     }
 }
 
-// Convert mrd::Calibration to ISMRMRD::MultibandCalibrationType
-ISMRMRD::MultibandCalibrationType convert(mrd::Calibration &m)
-{
-    if (m == mrd::Calibration::kSeparable2D)
-    {
-        return ISMRMRD::MultibandCalibrationType::SEPARABLE2D;
-    }
-    else if (m == mrd::Calibration::kFull3D)
-    {
-        return ISMRMRD::MultibandCalibrationType::FULL3D;
-    }
-    else if (m == mrd::Calibration::kOther)
-    {
-        return ISMRMRD::MultibandCalibrationType::OTHER;
-    }
-    else
-    {
-        throw std::runtime_error("Unknown Calibration");
-    }
-}
-
 // Convert ISMRMRD::Multiband to mrd::MultibandType
 mrd::MultibandType convert(ISMRMRD::Multiband &m)
 {
@@ -1259,21 +658,6 @@ mrd::MultibandType convert(ISMRMRD::Multiband &m)
     return multiband;
 }
 
-// Convert mrd::MultibandType to ISMRMRD::Multiband
-ISMRMRD::Multiband convert(mrd::MultibandType &m)
-{
-    ISMRMRD::Multiband multiband;
-    for (auto s : m.spacing)
-    {
-        multiband.spacing.push_back(convert(s));
-    }
-    multiband.deltaKz = m.delta_kz;
-    multiband.multiband_factor = m.multibandfactor;
-    multiband.calibration = convert(m.calibration);
-    multiband.calibration_encoding = m.calibrationencoding;
-    return multiband;
-}
-
 // Convert ISMRMRD::ParallelImaging to mrd::ParallelImagingType
 mrd::ParallelImagingType convert(ISMRMRD::ParallelImaging &p)
 {
@@ -1286,26 +670,6 @@ mrd::ParallelImagingType convert(ISMRMRD::ParallelImaging &p)
     if (p.interleavingDimension)
     {
         parallelImaging.interleaving_dimension = interleaving_dimension_from_string(*p.interleavingDimension);
-    }
-    if (p.multiband)
-    {
-        parallelImaging.multiband = convert(*p.multiband);
-    }
-    return parallelImaging;
-}
-
-// Convert mrd::ParallelImagingType to ISMRMRD::ParallelImaging
-ISMRMRD::ParallelImaging convert(mrd::ParallelImagingType &p)
-{
-    ISMRMRD::ParallelImaging parallelImaging;
-    parallelImaging.accelerationFactor = convert(p.acceleration_factor);
-    if (p.calibration_mode)
-    {
-        parallelImaging.calibrationMode = calibration_mode_to_string(*p.calibration_mode);
-    }
-    if (p.interleaving_dimension)
-    {
-        parallelImaging.interleavingDimension = interleaving_dimension_to_string(*p.interleaving_dimension);
     }
     if (p.multiband)
     {
@@ -1366,63 +730,6 @@ mrd::EncodingType convert(ISMRMRD::Encoding &e)
     if (e.echoTrainLength)
     {
         encoding.echo_train_length = *e.echoTrainLength;
-    }
-
-    return encoding;
-}
-
-// Convert mrd::EncodingType to ISMRMRD::Encoding
-ISMRMRD::Encoding convert(mrd::EncodingType &e)
-{
-    ISMRMRD::Encoding encoding;
-
-    encoding.encodedSpace = convert(e.encoded_space);
-    encoding.reconSpace = convert(e.recon_space);
-    encoding.encodingLimits = convert(e.encoding_limits);
-
-    if (e.trajectory == mrd::Trajectory::kCartesian)
-    {
-        encoding.trajectory = ISMRMRD::TrajectoryType::CARTESIAN;
-    }
-    else if (e.trajectory == mrd::Trajectory::kEpi)
-    {
-        encoding.trajectory = ISMRMRD::TrajectoryType::EPI;
-    }
-    else if (e.trajectory == mrd::Trajectory::kRadial)
-    {
-        encoding.trajectory = ISMRMRD::TrajectoryType::RADIAL;
-    }
-    else if (e.trajectory == mrd::Trajectory::kGoldenangle)
-    {
-        encoding.trajectory = ISMRMRD::TrajectoryType::GOLDENANGLE;
-    }
-    else if (e.trajectory == mrd::Trajectory::kSpiral)
-    {
-        encoding.trajectory = ISMRMRD::TrajectoryType::SPIRAL;
-    }
-    else if (e.trajectory == mrd::Trajectory::kOther)
-    {
-        encoding.trajectory = ISMRMRD::TrajectoryType::OTHER;
-    }
-    else
-    {
-        // throw error
-        throw std::runtime_error("Unknown trajectory type");
-    }
-
-    if (e.trajectory_description)
-    {
-        encoding.trajectoryDescription = convert(*e.trajectory_description);
-    }
-
-    if (e.parallel_imaging)
-    {
-        encoding.parallelImaging = convert(*e.parallel_imaging);
-    }
-
-    if (e.echo_train_length)
-    {
-        encoding.echoTrainLength = *e.echo_train_length;
     }
 
     return encoding;
@@ -1493,85 +800,10 @@ mrd::DiffusionDimension convert(ISMRMRD::DiffusionDimension &diffusionDimension)
     }
 }
 
-// Convert mrd::DiffusionDimension to ISMRMRD::DiffusionDimension
-ISMRMRD::DiffusionDimension convert(mrd::DiffusionDimension &diffusionDimension)
-{
-    if (diffusionDimension == mrd::DiffusionDimension::kAverage)
-    {
-        return ISMRMRD::DiffusionDimension::AVERAGE;
-    }
-    else if (diffusionDimension == mrd::DiffusionDimension::kContrast)
-    {
-        return ISMRMRD::DiffusionDimension::CONTRAST;
-    }
-    else if (diffusionDimension == mrd::DiffusionDimension::kPhase)
-    {
-        return ISMRMRD::DiffusionDimension::PHASE;
-    }
-    else if (diffusionDimension == mrd::DiffusionDimension::kRepetition)
-    {
-        return ISMRMRD::DiffusionDimension::REPETITION;
-    }
-    else if (diffusionDimension == mrd::DiffusionDimension::kSegment)
-    {
-        return ISMRMRD::DiffusionDimension::SEGMENT;
-    }
-    else if (diffusionDimension == mrd::DiffusionDimension::kSet)
-    {
-        return ISMRMRD::DiffusionDimension::SET;
-    }
-    else if (diffusionDimension == mrd::DiffusionDimension::kUser0)
-    {
-        return ISMRMRD::DiffusionDimension::USER_0;
-    }
-    else if (diffusionDimension == mrd::DiffusionDimension::kUser1)
-    {
-        return ISMRMRD::DiffusionDimension::USER_1;
-    }
-    else if (diffusionDimension == mrd::DiffusionDimension::kUser2)
-    {
-        return ISMRMRD::DiffusionDimension::USER_2;
-    }
-    else if (diffusionDimension == mrd::DiffusionDimension::kUser3)
-    {
-        return ISMRMRD::DiffusionDimension::USER_3;
-    }
-    else if (diffusionDimension == mrd::DiffusionDimension::kUser4)
-    {
-        return ISMRMRD::DiffusionDimension::USER_4;
-    }
-    else if (diffusionDimension == mrd::DiffusionDimension::kUser5)
-    {
-        return ISMRMRD::DiffusionDimension::USER_5;
-    }
-    else if (diffusionDimension == mrd::DiffusionDimension::kUser6)
-    {
-        return ISMRMRD::DiffusionDimension::USER_6;
-    }
-    else if (diffusionDimension == mrd::DiffusionDimension::kUser7)
-    {
-        return ISMRMRD::DiffusionDimension::USER_7;
-    }
-    else
-    {
-        throw std::runtime_error("Unknown DiffusionDimension");
-    }
-}
-
 // Convert ISMRMRD::GradientDirection to mrd::GradientDirectionType
 mrd::GradientDirectionType convert(ISMRMRD::GradientDirection &g)
 {
     mrd::GradientDirectionType gradientDirection;
-    gradientDirection.rl = g.rl;
-    gradientDirection.ap = g.ap;
-    gradientDirection.fh = g.fh;
-    return gradientDirection;
-}
-
-// Convert mrd::GradientDirectionType to ISMRMRD::GradientDirection
-ISMRMRD::GradientDirection convert(mrd::GradientDirectionType &g)
-{
-    ISMRMRD::GradientDirection gradientDirection;
     gradientDirection.rl = g.rl;
     gradientDirection.ap = g.ap;
     gradientDirection.fh = g.fh;
@@ -1583,15 +815,6 @@ mrd::DiffusionType convert(ISMRMRD::Diffusion &d)
 {
     mrd::DiffusionType diffusion;
     diffusion.gradient_direction = convert(d.gradientDirection);
-    diffusion.bvalue = d.bvalue;
-    return diffusion;
-}
-
-// Convert mrd::DiffusionType to ISMRMRD::Diffusion
-ISMRMRD::Diffusion convert(mrd::DiffusionType &d)
-{
-    ISMRMRD::Diffusion diffusion;
-    diffusion.gradientDirection = convert(d.gradient_direction);
     diffusion.bvalue = d.bvalue;
     return diffusion;
 }
@@ -1667,64 +890,6 @@ mrd::SequenceParametersType convert(ISMRMRD::SequenceParameters &s)
     return sequenceParameters;
 }
 
-// Convert mrd::SequenceParametersType to ISMRMRD::SequenceParameters
-ISMRMRD::SequenceParameters convert(mrd::SequenceParametersType &s)
-{
-    ISMRMRD::SequenceParameters sequenceParameters;
-
-    if (s.t_r.size())
-    {
-        sequenceParameters.TR = std::vector<float>(s.t_r.begin(), s.t_r.end());
-    }
-
-    if (s.t_e.size())
-    {
-        sequenceParameters.TE = std::vector<float>(s.t_e.begin(), s.t_e.end());
-    }
-
-    if (s.t_i.size())
-    {
-        sequenceParameters.TI = std::vector<float>(s.t_i.begin(), s.t_i.end());
-    }
-
-    if (s.flip_angledeg.size())
-    {
-        sequenceParameters.flipAngle_deg = std::vector<float>(s.flip_angledeg.begin(), s.flip_angledeg.end());
-    }
-
-    if (s.sequencetype)
-    {
-        sequenceParameters.sequence_type = *s.sequencetype;
-    }
-
-    if (s.echospacing.size())
-    {
-        sequenceParameters.echo_spacing = std::vector<float>(s.echospacing.begin(), s.echospacing.end());
-    }
-
-    if (s.diffusion_dimension)
-    {
-        sequenceParameters.diffusionDimension = convert(*s.diffusion_dimension);
-    }
-
-    if (s.diffusion.size())
-    {
-        std::vector<ISMRMRD::Diffusion> diffusion;
-        for (auto &d : s.diffusion)
-        {
-            diffusion.push_back(convert(d));
-        }
-        sequenceParameters.diffusion = diffusion;
-    }
-
-    if (s.diffusion_scheme)
-    {
-        sequenceParameters.diffusionScheme = *s.diffusion_scheme;
-    }
-
-    return sequenceParameters;
-}
-
 // Convert ISMRMRD::UserParameterString to mrd::UserParameterBase64Type
 mrd::UserParameterBase64Type convert_userbase64(ISMRMRD::UserParameterString &u)
 {
@@ -1732,15 +897,6 @@ mrd::UserParameterBase64Type convert_userbase64(ISMRMRD::UserParameterString &u)
     userParameterBase64.name = u.name;
     userParameterBase64.value = u.value;
     return userParameterBase64;
-}
-
-// Convert mrd::UserParameterBase64Type to ISMRMRD::UserParameterString
-ISMRMRD::UserParameterString convert_userbase64(mrd::UserParameterBase64Type &u)
-{
-    ISMRMRD::UserParameterString userParameterString;
-    userParameterString.name = u.name;
-    userParameterString.value = u.value;
-    return userParameterString;
 }
 
 // Convert ISMRMRD::UserParameters to mrd::UserParametersType
@@ -1766,34 +922,6 @@ mrd::UserParametersType convert(ISMRMRD::UserParameters &u)
     for (auto &p : u.userParameterBase64)
     {
         userParameters.user_parameter_base64.push_back(convert_userbase64(p));
-    }
-
-    return userParameters;
-}
-
-// Convert mrd::UserParametersType to ISMRMRD::UserParameters
-ISMRMRD::UserParameters convert(mrd::UserParametersType &u)
-{
-    ISMRMRD::UserParameters userParameters;
-
-    for (auto &p : u.user_parameter_long)
-    {
-        userParameters.userParameterLong.push_back(convert(p));
-    }
-
-    for (auto &p : u.user_parameter_double)
-    {
-        userParameters.userParameterDouble.push_back(convert(p));
-    }
-
-    for (auto &p : u.user_parameter_string)
-    {
-        userParameters.userParameterString.push_back(convert(p));
-    }
-
-    for (auto &p : u.user_parameter_base64)
-    {
-        userParameters.userParameterBase64.push_back(convert_userbase64(p));
     }
 
     return userParameters;
@@ -1832,39 +960,6 @@ mrd::WaveformType convert(ISMRMRD::WaveformType &w)
     }
 }
 
-// Convert mrd::WaveformType to ISMRMRD::WaveformType
-ISMRMRD::WaveformType convert(mrd::WaveformType &w)
-{
-    if (w == mrd::WaveformType::kEcg)
-    {
-        return ISMRMRD::WaveformType::ECG;
-    }
-    else if (w == mrd::WaveformType::kPulse)
-    {
-        return ISMRMRD::WaveformType::PULSE;
-    }
-    else if (w == mrd::WaveformType::kRespiratory)
-    {
-        return ISMRMRD::WaveformType::RESPIRATORY;
-    }
-    else if (w == mrd::WaveformType::kTrigger)
-    {
-        return ISMRMRD::WaveformType::TRIGGER;
-    }
-    else if (w == mrd::WaveformType::kGradientwaveform)
-    {
-        return ISMRMRD::WaveformType::GRADIENTWAVEFORM;
-    }
-    else if (w == mrd::WaveformType::kOther)
-    {
-        return ISMRMRD::WaveformType::OTHER;
-    }
-    else
-    {
-        throw std::runtime_error("Unknown WaveformType");
-    }
-}
-
 // Convert ISMRMRD::WaveformInformation to mrd::WaveformInformationType
 mrd::WaveformInformationType convert(ISMRMRD::WaveformInformation &w)
 {
@@ -1875,16 +970,6 @@ mrd::WaveformInformationType convert(ISMRMRD::WaveformInformation &w)
     {
         waveformInformation.user_parameters = convert(*w.userParameters);
     }
-    return waveformInformation;
-}
-
-// Convert mrd::WaveformInformationType to ISMRMRD::WaveformInformation
-ISMRMRD::WaveformInformation convert(mrd::WaveformInformationType &w)
-{
-    ISMRMRD::WaveformInformation waveformInformation;
-    waveformInformation.waveformName = w.waveform_name;
-    waveformInformation.waveformType = convert(w.waveform_type);
-    waveformInformation.userParameters = convert(w.user_parameters);
     return waveformInformation;
 }
 
@@ -1950,68 +1035,6 @@ mrd::Header convert(ISMRMRD::IsmrmrdHeader &hdr)
     return h;
 }
 
-// Convert mrd::Header to ISMRMRD::IsmrmrdHeader
-ISMRMRD::IsmrmrdHeader convert(mrd::Header &hdr)
-{
-    ISMRMRD::IsmrmrdHeader h;
-
-    if (hdr.version)
-    {
-        h.version = *hdr.version;
-    }
-
-    if (hdr.subject_information)
-    {
-        h.subjectInformation = convert(*hdr.subject_information);
-    }
-
-    if (hdr.study_information)
-    {
-        h.studyInformation = convert(*hdr.study_information);
-    }
-
-    if (hdr.measurement_information)
-    {
-        h.measurementInformation = convert(*hdr.measurement_information);
-    }
-
-    if (hdr.acquisition_system_information)
-    {
-        h.acquisitionSystemInformation = convert(*hdr.acquisition_system_information);
-    }
-
-    h.experimentalConditions = convert(hdr.experimental_conditions);
-
-    if (hdr.encoding.size() > 0)
-    {
-        for (auto e : hdr.encoding)
-        {
-            h.encoding.push_back(convert(e));
-        }
-    }
-    else
-    {
-        throw std::runtime_error("No encoding found in mrd header");
-    }
-
-    if (hdr.sequence_parameters)
-    {
-        h.sequenceParameters = convert(*hdr.sequence_parameters);
-    }
-
-    if (hdr.user_parameters)
-    {
-        h.userParameters = convert(*hdr.user_parameters);
-    }
-
-    for (auto w : hdr.waveform_information)
-    {
-        h.waveformInformation.push_back(convert(w));
-    }
-
-    return h;
-}
-
 // Convert ISMRMRD::EncodingCounters to mrd::EncodingCounters
 mrd::EncodingCounters convert(ISMRMRD::EncodingCounters &e)
 {
@@ -2028,110 +1051,6 @@ mrd::EncodingCounters convert(ISMRMRD::EncodingCounters &e)
     for (auto &u : e.user)
     {
         encodingCounters.user.push_back(u);
-    }
-
-    return encodingCounters;
-}
-
-// Convert mrd::EncodingCounters to ISMRMRD::EncodingCounters
-ISMRMRD::EncodingCounters convert(mrd::EncodingCounters &e)
-{
-    ISMRMRD::EncodingCounters encodingCounters;
-    if (e.kspace_encode_step1)
-    {
-        encodingCounters.kspace_encode_step_1 = *e.kspace_encode_step1;
-    }
-    else
-    {
-        encodingCounters.kspace_encode_step_1 = 0;
-    }
-
-    if (e.kspace_encode_step2)
-    {
-        encodingCounters.kspace_encode_step_2 = *e.kspace_encode_step2;
-    }
-    else
-    {
-        encodingCounters.kspace_encode_step_2 = 0;
-    }
-
-    if (e.average)
-    {
-        encodingCounters.average = *e.average;
-    }
-    else
-    {
-        encodingCounters.average = 0;
-    }
-
-    if (e.slice)
-    {
-        encodingCounters.slice = *e.slice;
-    }
-    else
-    {
-        encodingCounters.slice = 0;
-    }
-
-    if (e.contrast)
-    {
-        encodingCounters.contrast = *e.contrast;
-    }
-    else
-    {
-        encodingCounters.contrast = 0;
-    }
-
-    if (e.phase)
-    {
-        encodingCounters.phase = *e.phase;
-    }
-    else
-    {
-        encodingCounters.phase = 0;
-    }
-
-    if (e.repetition)
-    {
-        encodingCounters.repetition = *e.repetition;
-    }
-    else
-    {
-        encodingCounters.repetition = 0;
-    }
-
-    if (e.set)
-    {
-        encodingCounters.set = *e.set;
-    }
-    else
-    {
-        encodingCounters.set = 0;
-    }
-
-    if (e.segment)
-    {
-        encodingCounters.segment = *e.segment;
-    }
-    else
-    {
-        encodingCounters.segment = 0;
-    }
-
-    if (e.user.size() > 8)
-    {
-        throw std::runtime_error("Too many user encoding counters");
-    }
-
-    int i = 0;
-    for (auto &u : e.user)
-    {
-        // u has to fit into a 16 bit unsigned integer
-        if (u > 65535)
-        {
-            throw std::runtime_error("User encoding counter too large");
-        }
-        encodingCounters.user[i++] = static_cast<uint16_t>(u);
     }
 
     return encodingCounters;
@@ -2216,6 +1135,7 @@ mrd::Acquisition convert(ISMRMRD::Acquisition &acq)
     return acquisition;
 }
 
+// Convert ISMRMRD::Waveform to mrd::Waveform
 mrd::Waveform<uint32_t> convert(ISMRMRD::Waveform &wfm)
 {
     mrd::Waveform<uint32_t> waveform;
@@ -2238,6 +1158,7 @@ mrd::Waveform<uint32_t> convert(ISMRMRD::Waveform &wfm)
     return waveform;
 }
 
+// Convert mrd::Image<T> to ISMRMRD::Image<T>
 template <typename T>
 mrd::Image<T> convert(ISMRMRD::Image<T> &im)
 {
